requirements.txt
opencv-python>=4.8
numpy>=1.24
Pillow>=10.0
streamlit>=1.30
